Refactor metric recorder registration
