+++ src/fairseq2/metrics/recorders/_tensorboard.py
@@ -13,6 +13,7 @@ from typing import Final, final
 
 from typing_extensions import override
 
+from fairseq2.context import RuntimeContext
 from fairseq2.logging import log
 from fairseq2.metrics import MetricDescriptor
 from fairseq2.metrics.recorders._handler import MetricRecorderHandler
@@ -142,3 +143,13 @@ class TensorBoardRecorderHandler(MetricRecorderHandler):
     @override
     def config_kls(self) -> type[object]:
         return TensorBoardRecorderConfig
+
+
+def register_tensorboard_recorder(context: RuntimeContext) -> None:
+    metric_descriptors = context.get_registry(MetricDescriptor)
+
+    handler = TensorBoardRecorderHandler(metric_descriptors)
+
+    registry = context.get_registry(MetricRecorderHandler)
+
+    registry.register(TENSORBOARD_RECORDER, handler)
