+++ src/fairseq2/profilers/_torch.py
