+++ src/fairseq2/recipes/setup.py
@@ -9,18 +9,10 @@ from __future__ import annotations
 from fairseq2.context import RuntimeContext
 from fairseq2.recipes.asr import register_asr_eval_configs
 from fairseq2.recipes.lm import (
-    CPO_FINETUNE_UNIT,
-    DPO_FINETUNE_UNIT,
-    ORPO_FINETUNE_UNIT,
-    SIMPO_FINETUNE_UNIT,
-    CpoFinetuneUnitHandler,
-    DpoFinetuneUnitHandler,
-    OrpoFinetuneUnitHandler,
-    POFinetuneUnitHandler,
-    SimPOFinetuneUnitHandler,
     register_instruction_finetune_configs,
     register_lm_loss_eval_configs,
     register_po_finetune_configs,
+    register_po_finetune_units,
     register_text_generate_configs,
 )
 from fairseq2.recipes.mt import (
@@ -35,7 +27,7 @@ from fairseq2.recipes.wav2vec2 import (
 from fairseq2.recipes.wav2vec2.asr import register_wav2vec2_asr_train_configs
 
 
-def _register_recipes(context: RuntimeContext) -> None:
+def register_recipes(context: RuntimeContext) -> None:
     register_asr_eval_configs(context)
     register_instruction_finetune_configs(context)
     register_lm_loss_eval_configs(context)
@@ -48,13 +40,4 @@ def _register_recipes(context: RuntimeContext) -> None:
     register_wav2vec2_eval_configs(context)
     register_wav2vec2_train_configs(context)
 
-    _register_po_finetune_units(context)
-
-
-def _register_po_finetune_units(context: RuntimeContext) -> None:
-    registry = context.get_registry(POFinetuneUnitHandler)
-
-    registry.register(CPO_FINETUNE_UNIT, CpoFinetuneUnitHandler())
-    registry.register(DPO_FINETUNE_UNIT, DpoFinetuneUnitHandler(context))
-    registry.register(ORPO_FINETUNE_UNIT, OrpoFinetuneUnitHandler())
-    registry.register(SIMPO_FINETUNE_UNIT, SimPOFinetuneUnitHandler())
+    register_po_finetune_units(context)
